# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_handler.AddCartRequest:
    properties:
      group_buy_id:
        type: integer
      group_buy_team_id:
        type: integer
      product_id:
        type: integer
      purchase_type:
        example: 1
        type: integer
      quantity:
        example: 1
        type: integer
      spec:
        type: string
    required:
    - product_id
    type: object
...
  internal_handler.CreateOrderRequest:
    properties:
      activity_product_id:
        example: 1
        type: integer
      address_id:
        type: integer
      cart_item_id:
        type: integer
      delivery_type:
        example: 1
        type: integer
      group_buy_id:
        type: integer
      group_buy_team_id:
        type: integer
      merchant_id:
        type: integer
      product_id:
        example: 1
        type: integer
      purchase_type:
        example: 1
        type: integer
      quantity:
        example: 1
        type: integer
      remark:
        type: string
      user_coupon_id:
        example: 1
        type: integer
    required:
    - merchant_id
    type: object
...
  yujixinjiang_backend_internal_model.Account:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      gender:
        type: integer
      id:
        type: integer
      is_rider:
        type: integer
      last_login_at:
        type: string
      nickname:
        type: string
      openid:
        type: string
      phone:
        type: string
      status:
        type: integer
      type:
        type: integer
      unionid:
        type: string
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.AddressSnapshot:
    properties:
      city:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      detail:
        type: string
      district:
        type: string
      province:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.CartItem:
    properties:
      account_id:
        type: integer
      created_at:
        type: string
      group_buy_id:
        type: integer
      group_buy_team_id:
        type: integer
      id:
        type: integer
      product:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Product'
      product_id:
        type: integer
      purchase_type:
        type: integer
      quantity:
        type: integer
      selected:
        type: integer
      spec:
        type: string
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.MerchantProfile:
    properties:
      account:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Account'
      account_id:
        type: integer
      address:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      shop_logo:
        type: string
      shop_name:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.OrderItem:
    properties:
      activity_id:
        type: integer
      activity_product_id:
        type: integer
      created_at:
        type: string
      group_buy_id:
        type: integer
      group_buy_team_id:
        type: integer
      id:
        type: integer
      order_id:
        type: integer
      product_id:
        type: integer
      product_image:
        type: string
      product_name:
        type: string
      purchase_type:
        type: integer
      quantity:
        type: integer
      spec:
        type: string
      subtotal:
        type: number
      unit_price:
        type: number
    type: object
  yujixinjiang_backend_internal_model.Product:
    properties:
      category:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.ProductCategory'
      category_id:
        type: integer
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      enable_coupon:
        type: integer
      enable_group_buy:
        type: integer
      group_buy_allow_repeat:
        type: integer
      group_buy_price:
        type: number
      group_buy_target_count:
        type: integer
      id:
        type: integer
      images:
        items:
          type: string
        type: array
      is_hot:
        type: integer
      item_type:
        type: integer
      merchant:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.MerchantProfile'
      merchant_id:
        type: integer
      name:
        type: string
      original_price:
        type: number
      price:
        type: number
      sales_count:
        type: integer
      status:
        type: integer
      stock:
        type: integer
      updated_at:
        type: string
    type: object
  yujixinjiang_backend_internal_model.ProductCategory:
    properties:
      created_at:
        type: string
      icon_url:
        type: string
      id:
        type: integer
      merchant_id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
      status:
        type: integer
      updated_at:
        type: string
    type: object
  yujixinjiang_backend_internal_model.UserAddress:
    properties:
      account_id:
        type: integer
      city:
        type: string
      contact_name:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      detail:
        type: string
      district:
        type: string
      id:
        type: integer
      is_default:
        type: integer
      province:
        type: string
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_model.UserProfile:
    properties:
      account_id:
        type: integer
      bio:
        type: string
      birthday:
        type: string
      city:
        type: string
      created_at:
        type: string
      district:
        type: string
      id:
        type: integer
      points:
        type: integer
      province:
        type: string
      real_name:
        type: string
      updated_at:
        type: string
    type: object
  yujixinjiang_backend_internal_query.PageResult:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  yujixinjiang_backend_internal_response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  yujixinjiang_backend_internal_service.BuyerBrief:
    properties:
      account_id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
    type: object
  yujixinjiang_backend_internal_service.CartItemView:
    properties:
      account_id:
        type: integer
      can_group_buy:
        type: boolean
      can_use_coupon:
        type: boolean
      created_at:
        type: string
      group_buy_id:
        type: integer
      group_buy_team_id:
        type: integer
      group_price:
        type: number
      id:
        type: integer
      product:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Product'
      product_id:
        type: integer
      purchase_type:
        type: integer
      quantity:
        type: integer
      selected:
        type: integer
      spec:
        type: string
      subtotal:
        type: number
      unit_price:
        type: number
      updated_at:
        type: string
    type: object
...
  yujixinjiang_backend_internal_service.GroupBuyProgress:
    properties:
      allow_repeat_join:
        type: integer
      current_count:
        type: integer
      expire_at:
        type: string
      group_price:
        type: number
      is_leader:
        type: boolean
      remaining_count:
        type: integer
      status:
        type: integer
      status_text:
        type: string
      target_count:
        type: integer
      team_id:
        type: integer
      user_join_count:
        type: integer
      user_joined:
        type: boolean
    type: object
...
  yujixinjiang_backend_internal_service.OrderView:
    properties:
      account_id:
        type: integer
      activity_id:
        type: integer
      address_snapshot:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.AddressSnapshot'
      buyer:
        $ref: '#/definitions/yujixinjiang_backend_internal_service.BuyerBrief'
      created_at:
        type: string
      delivery_type:
        type: integer
      discount_amount:
        type: number
      group_buy_progress:
        $ref: '#/definitions/yujixinjiang_backend_internal_service.GroupBuyProgress'
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/yujixinjiang_backend_internal_model.OrderItem'
        type: array
      merchant_id:
        type: integer
      merchant_review_stage:
        type: integer
      order_no:
        type: string
      paid_at:
        type: string
      pay_amount:
        type: number
      pay_status:
        type: integer
      remark:
        type: string
      status:
        type: integer
      status_code:
        type: string
      status_text:
        type: string
      total_amount:
        type: number
      updated_at:
        type: string
      user_coupon_id:
        type: integer
      verify_code:
        type: string
    type: object
  yujixinjiang_backend_internal_service.OverviewResponse:
    properties:
      account:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Account'
      profile:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.UserProfile'
      stats:
        $ref: '#/definitions/yujixinjiang_backend_internal_service.UserStats'
    type: object
...
  yujixinjiang_backend_internal_service.ProfileDetail:
    properties:
      account:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.Account'
      addresses:
        items:
          $ref: '#/definitions/yujixinjiang_backend_internal_model.UserAddress'
        type: array
      profile:
        $ref: '#/definitions/yujixinjiang_backend_internal_model.UserProfile'
      stats:
        $ref: '#/definitions/yujixinjiang_backend_internal_service.UserStats'
    type: object
...
  yujixinjiang_backend_internal_service.UserStats:
    properties:
      address_count:
        type: integer
      cart_count:
        type: integer
      coupon_unused_count:
        type: integer
      inventory_kind_count:
        type: integer
      inventory_total_qty:
        type: integer
      order_count:
        type: integer
    type: object
...
info:
  contact:
    email: support@yujixinjiang.com
    name: API Support
  description: 豫记信疆微信小程序后端接口文档，支持在线调试。
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 豫记信疆 API
  version: "1.0"
paths:
...
  /user/cart:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/yujixinjiang_backend_internal_service.CartItemView'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 购物车
      tags:
      - 用户
    post:
      consumes:
      - application/json
      parameters:
      - description: 加购信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.AddCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_model.CartItem'
              type: object
      security:
      - BearerAuth: []
      summary: 加入购物车
      tags:
      - 用户-购物车
...
  /user/coupons:
    get:
      parameters:
      - description: 优惠券状态，0=未使用
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 优惠券列表
      tags:
      - 用户
...
  /user/inventory:
    get:
      description: 购买成功后商品会进入背包；使用 POST /user/inventory/{id}/use
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 背包库存
      tags:
      - 用户
...
  /user/orders:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: pending_group|pending_merchant|approved|...
        in: query
        name: status_code
        type: string
      - description: 订单状态筛选
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_query.PageResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 历史订单列表
      tags:
      - 用户
    post:
      consumes:
      - application/json
      description: 暂无支付流程，创建后 pay_status=已支付。可选 user_coupon_id 使用优惠券
      parameters:
      - description: 下单信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.OrderView'
              type: object
      security:
      - BearerAuth: []
      summary: 创建订单
      tags:
      - 用户-订单
  /user/orders/{id}:
    get:
      parameters:
      - description: 订单 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.OrderView'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 订单详情
      tags:
      - 用户
...
  /user/overview:
    get:
      description: 返回资料摘要与各项数量统计
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.OverviewResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 个人中心概览
      tags:
      - 用户
  /user/profile:
    get:
      description: 返回账号、扩展资料、地址列表与统计
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
            - properties:
                data:
                  $ref: '#/definitions/yujixinjiang_backend_internal_service.ProfileDetail'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/yujixinjiang_backend_internal_response.Body'
      security:
      - BearerAuth: []
      summary: 完整个人信息
      tags:
      - 用户
...
securityDefinitions:
  BearerAuth:
    description: '登录后获取的 JWT，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"